refactor: extract PDF semantic flow boundaries
